+++ src/TiempoInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace TrabajoSube;
+
+interface TiempoInterface {
+    public function time();
+
+}
+?>
