+++ src/TiempoReal.php
@@ -1,10 +1,6 @@
 <?php
 namespace TrabajoSube;
 
-interface TiempoInterface {
-    public function time();
-
-}
 
 class TiempoReal implements TiempoInterface 
 {
@@ -13,21 +9,5 @@ class TiempoReal implements TiempoInterface
     }
 }
 
-class TiempoFalso implements TiempoInterface 
-{
-    protected $tiempo; 
-    public function __construct($iniciarEn = 0){
-        $this->tiempo = $iniciarEn;
-    }
-
-    public function time(){
-        return $this->tiempo;
-    }
-
-    public function avanzar($segundos){
-        $this->tiempo += $segundos;
-    }
-
-}
 
 ?>
